Remove -display argument from vncviewer
It has been broken for some time anyway, as FLTK opens the display
behind our backs and before we've had a time to check the arguments.

Rather than fixing this, let's remove it and avoid the complexity. Users
can still use the environment variable for a more reliable way of
controlling the X11 server used.