Open about dialog asynchronously
Stop creating a problematic nested loop and instead return to the
normal mainloop.

This required adding a new callback mechanism for the dialogs. The
normal callback is only fired when something external closes the dialog,
not when it is closed by ourselves (e.g. when clicking the ok button).

The memory for the dialog pointer isn't freed because we cannot
guarantee that the memory isn't accessed after finished() has been
called. Note that this results in a memory leak.