TEMP: Refactor file dialog handling
Removed usedDir to make the file dialog default to opening in the same
directory on each reopen, to simplify code and align with standard
behavior on other applications. Renamed file_chooser to fileChooser
align with the naming conventions used in the qtleft branch. Also
updated handleLoad and handleSaveAs to pass the filedialog instead of
only the filename for consistency with qtleft.