Reinitialize context menu after options dialog
To align the functionality with its counterpart on the fltk branch, the
context menu is now reinitialized after the options dialog has been
opened. Added the functions setMenuKey and handleOptions to reinitialize
the context menu with the updated settings.

Note that the context menu cannot be toggled immediately after logging
in - you need to double-click in the window first. However, this was the
case previously and is not a regression.